PG and SQLite share search impl
