+++ db-service/lib/cql-functions.js
@@ -1,7 +1,28 @@
-const cds = require("@sap/cds")
+const cds = require('@sap/cds')
 
 // OData: https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_CanonicalFunctions
 const StandardFunctions = {
+  /**
+   * Generates SQL statement that produces a boolean value indicating whether the search term is contained in the given columns
+   * @param {string} ref
+   * @param {string} arg
+   * @returns {string}
+   */
+  search: function (ref, arg) {
+    if (!('val' in arg)) throw new Error(`Only single value arguments are allowed for $search`)
+    // only apply first search term, rest is ignored
+    const sub = /("")|("(?:[^"]|\\")*(?:[^\\]|\\\\)")|(\S*)/.exec(arg.val)
+    let val
+    try {
+      val = (sub[2] ? JSON.parse(sub[2]) : sub[3]) || ''
+    } catch {
+      val = sub[2] || sub[3] || ''
+    }
+    arg.val = arg.__proto__.val = val
+    const refs = ref.list
+    const { toString } = ref
+    return '(' + refs.map(ref2 => this.contains(this.tolower(toString(ref2)), this.tolower(arg))).join(' or ') + ')'
+  },
   /**
    * Generates SQL statement that produces the length of a given string
    * @param {string} x
