+++ sqlite/lib/cql-functions.js
@@ -7,27 +7,6 @@ const StandardFunctions = {
   current_date: () => 'current_date',
   current_time: () => 'current_time',
   /**
-   * Generates SQL statement that produces a boolean value indicating whether the search term is contained in the given columns
-   * @param {string} ref
-   * @param {string} arg
-   * @returns {string}
-   */
-  search: function (ref, arg) {
-    if (!('val' in arg)) throw new Error(`Only single value arguments are allowed for $search`)
-    // only apply first search term, rest is ignored
-    const sub = /("")|("(?:[^"]|\\")*(?:[^\\]|\\\\)")|(\S*)/.exec(arg.val)
-    let val
-    try {
-      val = (sub[2] ? JSON.parse(sub[2]) : sub[3]) || ''
-    } catch {
-      val = sub[2] || sub[3] || ''
-    }
-    arg.val = arg.__proto__.val = val
-    const refs = ref.list
-    const { toString } = ref
-    return '(' + refs.map(ref2 => this.contains(this.tolower(toString(ref2)), this.tolower(arg))).join(' or ') + ')'
-  },
-  /**
    * Generates SQL statement that produces a boolean value indicating whether the first string contains the second string
    * @param  {...string} args
    * @returns {string}
